Cleanup, renamed PhysicalMMU into ByteMMU to better reflect the byte-addressable property.

+++ PCSX2_Core/Source/Common/Util/MathUtil/MathUtil.cpp
@@ -28,7 +28,7 @@ u32 MathUtil::countLeadingBits(s32 value)
 	return num_leading_bits;
 }
 
-u32 MathUtil::constructMask32(u8 startPos, u8 length)
+u32 MathUtil::constructMask32(int startPos, int length)
 {
 	// Due to how x86 (and possibly other arch's) shift, we need a solution for the case when length == 32, as it generates a bad mask.
 	// Old method: mask = (1 << param) - 1;
@@ -36,20 +36,20 @@ u32 MathUtil::constructMask32(u8 startPos, u8 length)
 	return (static_cast<u32>(-(length != 0)) & (static_cast<u32>(-1) >> (32 - length))) << startPos;
 }
 
-u32 MathUtil::extractMaskedValue32(u32 value, u8 maskStartPos, u8 maskLength)
+u32 MathUtil::extractMaskedValue32(u32 value, int maskStartPos, int maskLength)
 {
 	const u32 mask = constructMask32(maskStartPos, maskLength);
 	return ((value & mask) >> maskStartPos);
 }
 
-u32 MathUtil::insertMaskedValue32(u32 value, u32 insertValue, u8 maskStartPos, u8 maskLength)
+u32 MathUtil::insertMaskedValue32(u32 value, u32 insertValue, int maskStartPos, int maskLength)
 {
 	const u32 mask = constructMask32(maskStartPos, maskLength);
 	const u32 insertValueShifted = (insertValue << maskStartPos) & mask;
 	return ((value & ~mask) | insertValueShifted);
 }
 
-u16 MathUtil::constructMask16(u8 startPos, u8 length)
+u16 MathUtil::constructMask16(int startPos, int length)
 {
 	// Due to how x86 (and possibly other arch's) shift, we need a solution for the case when length == 16, as it generates a bad mask.
 	// Old method: mask = (1 << param) - 1;
@@ -57,13 +57,13 @@ u16 MathUtil::constructMask16(u8 startPos, u8 length)
 	return (static_cast<u16>(-(length != 0)) & (static_cast<u16>(-1) >> (16 - length))) << startPos;
 }
 
-u16 MathUtil::extractMaskedValue16(u16 value, u8 maskStartPos, u8 maskLength)
+u16 MathUtil::extractMaskedValue16(u16 value, int maskStartPos, int maskLength)
 {
 	const u16 mask = constructMask16(maskStartPos, maskLength);
 	return ((value & mask) >> maskStartPos);
 }
 
-u16 MathUtil::insertMaskedValue16(u16 value, u16 insertValue, u8 maskStartPos, u8 maskLength)
+u16 MathUtil::insertMaskedValue16(u16 value, u16 insertValue, int maskStartPos, int maskLength)
 {
 	const u16 mask = constructMask16(maskStartPos, maskLength);
 	const u16 insertValueShifted = (insertValue << maskStartPos) & mask;
